10.3389/fevo.2021.762173/requirements.txt
coloredlogs==15.0
cycler==0.10.0
decorator==4.4.2
geographiclib==1.50
geopy==2.1.0
humanfriendly==9.1
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.2
six==1.15.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
